% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaplot.R
\name{pcaplot}
\alias{pcaplot}
\title{pca}
\usage{
pcaplot(mat, title = "PCA Plot", col = rownames(mat))
}
\arguments{
\item{mat}{(matrix/dataframe):  mat}

\item{title}{(character) :}

\item{subtype}{(dataframe metadata) : rownames correspnond to}

\item{labe}{(character) : rownames correspnond to}
}
\value{
ggplot
}
\description{
pca
}
