% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_log_exp.R
\name{rel_log_exp}
\alias{rel_log_exp}
\title{relative log expression calculation}
\usage{
rel_log_exp(Y)
}
\arguments{
\item{Y}{data matrix. Rows are observations and columns are features (e.g. genes).}
}
\value{
numeric
}
\description{
relative log expression calculation
}
